Refactored and moved color components into the fxcomponents package
