+++ TripoliApp/src/main/java/org/cirdles/tripoli/gui/settings/color/SpeciesColorRowSelectionRecord.java
@@ -1,5 +1,7 @@
 package org.cirdles.tripoli.gui.settings.color;
 
+import org.cirdles.tripoli.gui.settings.color.fxcomponents.SpeciesColorPane;
+import org.cirdles.tripoli.gui.settings.color.fxcomponents.SpeciesColorRow;
 import org.cirdles.tripoli.species.SpeciesColorSetting;
 
 public record SpeciesColorRowSelectionRecord(SpeciesColorPane speciesColorPane,
